unit tests

+++ src/main/java/com/czertainly/core/model/cbom/BomSearchRequestDto.java
@@ -3,7 +3,6 @@ package com.czertainly.core.model.cbom;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.Setter;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 import jakarta.validation.constraints.NotNull;
@@ -19,7 +18,7 @@ public class BomSearchRequestDto {
                 example = "1769156084",
                 requiredMode = Schema.RequiredMode.REQUIRED
         )
-        private Integer after;
+        private Long after;
 
     @Override
     public String toString() {
